move nav to own component
